+++ tests/testthat/test-time.R
@@ -66,6 +66,7 @@ test_with_dir("build time the same after superfluous make", {
 test_with_dir("runtime predictions", {
   skip_on_cran() # CRAN gets whitelist tests only (check time limits).
   skip_if_not_installed("lubridate")
+  skip_if_not_installed("txtq")
   con <- dbug()
   expect_warning(p0 <- as.numeric(predict_runtime(con)))
   expect_true(p0 < 1e4)