+++ NEWS.md
@@ -15,6 +15,9 @@ to tell the user if the command, a dependency, an input file, or an ouptut file 
 - Add a new `code_to_plan()` function to turn R scripts and R Markdown reports into workflow plan data frames.
 - Add a new `drake_plan_source()` function, which generates lines of code for a `drake_plan()` call. This `drake_plan()` call produces the plan passed to `drake_plan_source()`. The main purpose is visual inspection (we even have syntax highlighting via `prettycode`) but users may also save the output to a script file for the sake of reproducibility or simple reference.
 - Memoize all the steps of `build_drake_graph()` and print to the console the ones that execute.
+- Bug fix: avoid `sort(NULL)` because it returns `NA`s on R 3.3.
+- Bug fix: work around mysterious `codetools` failures on R 3.3 (add a `tryCatch()` statement in `find_globals()`).
+- Skip some tests if `txtq` is not installed.
 
 # Version 5.4.0